+++ src/components/ServiceDetails/AboutService/ServiceDemo/index.js
@@ -231,7 +231,7 @@ const mapDispatchToProps = (dispatch, ownProps) => ({
   fetchMeteringData: args => dispatch(serviceDetailsActions.fetchMeteringData(args)),
   startWalletDetailsPolling: (orgId, groupId) => dispatch(userActions.startWalletDetailsPolling(orgId, groupId)),
   stopWalletDetailsPolling: () => dispatch(userActions.stopWalletDetailsPolling),
-  cancelOrder: () => dispatch(userActions.cancelOrder),
+  cancelOrder: orderId => dispatch(paymentActions.cancelOrder(orderId)),
   fetchOrderDetails: orderId => dispatch(paymentActions.fetchOrderDetails(orderId)),
   updateWallet: walletDetails => dispatch(userActions.updateWallet(walletDetails)),
   updatePaypalInProgress: (orderId, orderType, paymentId, paypalPaymentId, PayerID) =>
